api-service/requirements.txt
# my_project/api-service/requirements.txt
Flask
mysql-connector-python